minor cleanup, move apiGen code to ApiGen.js since we will have more code added to this module for api-impl-regen etc..
